move `Root` to submodule
